+++ testing/src/main/java/Playground.java
@@ -1,23 +1,38 @@
-public class Playground {
-  public static void main(String[] args) throws InterruptedException {
-    int[] buffer = new int[10];
+import dev.gamekit.core.Application;
+import dev.gamekit.core.IO;
+import dev.gamekit.core.Renderer;
+import dev.gamekit.core.Scene;
+import dev.gamekit.settings.Resolution;
+import dev.gamekit.settings.Settings;
+import dev.gamekit.utils.Picture;
 
-    for (int i = -10; i < 0; i++) {
-      System.out.printf("%d => %d%n", i, getWrappedIndex(buffer, i));
-    }
+import java.awt.*;
+
+public class Playground extends Scene {
+  private static final Picture SIMPLE = IO.getImage("keep-it-simple-light.jpg");
+  private static final Picture PLANET_FALL = IO.getImage("planetfall-artwork.jpg");
+  private static final Picture MASK = IO.getImage("planetfall-logo-mask.png");
+
+  public Playground() {
+    super("Playground");
   }
 
-  protected static int getWrappedIndex(int[] buffer, int index) {
-    int newIndex = index;
+  public static void main(String[] args) {
+    Application game = new Application(
+      new Settings(
+        "Playground",
+        Resolution.FULL_HD
+      )
+    ) { };
+    game.loadScene(new Playground());
+    game.run();
+  }
 
-    if (newIndex >= buffer.length) {
-      while (newIndex >= buffer.length)
-        newIndex -= buffer.length;
-    } else if (newIndex < 0) {
-      while (newIndex < 0)
-        newIndex += buffer.length;
-    }
-
-    return newIndex;
+  @Override
+  public void render() {
+    Renderer.clear(Color.BLACK);
+    Renderer.drawImage(SIMPLE, 0, 0, SIMPLE.getWidth(), SIMPLE.getHeight());
+    Renderer.drawImage(PLANET_FALL, 0, 0, PLANET_FALL.getWidth(), PLANET_FALL.getHeight())
+      .withMaskImage(MASK);
   }
 }